Design Improvements to DS Dropdowns
