+++ src/components/fileHandling/UploadFile.css
@@ -25,7 +25,7 @@
   display: flex;
   flex-wrap: wrap;
   align-items: center;
-  gap: 10px;
+  gap: 20px;
   padding: 10px;
   /*background-color: #f4f4f4;*/
   background: transparent !important;
@@ -438,7 +438,7 @@
 
 /* Header for each section in the Descriptive Statistics Panel */
 .statistics-panel h4 {
-  background-color: #f1f1f1;
+  background-color: #e7e7e7;
   color: #333;
   padding: 10px;
   margin: 0;
@@ -452,30 +452,46 @@
 }
 
 /* Risk level categories for color */
-.sevLvl {
+.dsSevereLevelCard {
   background-color: #cb0033a8;
   /* red */
 }
+.dsSevLevelDrop {
+  background-color: #cb00332f;
+}
 
-.highLvl {
+.dsHighLevelCard {
   background-color: #ff66009f;
   /* orange */
 }
+.dsHighLevelDrop {
+  background-color: #ff660033;
+}
 
-.modLvl {
+.dsModerateLevelCard {
   background-color: #fde001a5;
   /* yellow */
 }
+.dsModLevelDrop {
+  background-color: #fde00133;
+}
 
-.minLvl {
+.dsMinorLevelCard {
   background-color: #3565cd95;
   /* light green */
 }
+.dsMinLevelDrop {
+  background-color: #3565cd34;
+}
 
-.insigLvl {
+.dsInsignificantLevelCard {
   background-color: #009a679b;
   /* green */
 }
+.dsInsigLevelDrop {
+  background-color: #009a671f;
+}
+
 
 .statistics-panel button {
   border:none;
@@ -491,14 +507,6 @@
   background-color: #00000031;
 }
 
-.dsRiskDropMenu {
-  position: absolute;
-  height: 300px;
-  width: 200px;
-  background-color: #ccc;
-  overflow:scroll;
-}
-
 /* Descriptive Statistics Button */
 .statistics-button {
   position: absolute;
@@ -510,10 +518,45 @@
   border: none;
   border-radius: 4px;
   margin-bottom: 10px;
+  margin-top: 10px;
+  margin-right: -150px;
 }
 
 .statistics-button:hover {
   background-color: #555;
+}
+
+.dsSevLevelDrop dt,
+.dsHighLevelDrop dt,
+.dsModLevelDrop dt,
+.dsMinLevelDrop dt,
+.dsInsigLevelDrop dt{  
+  font-size: 16px;
+  margin-left: 5px;
+  margin-top: 0px;
+  text-wrap: wrap;
+  word-wrap: break-word;
+}
+
+.dsSevLevelDrop dd,
+.dsHighLevelDrop dd,
+.dsModLevelDrop dd,
+.dsMinLevelDrop dd,
+.dsInsigLevelDrop dd{
+  display: list-item;
+  list-style-type: disc;
+  margin-left: 25px;
+  margin-bottom: 0px;
+  text-wrap: balance;
+
+}
+
+.dsSevLevelDrop p,
+.dsHighLevelDrop p,
+.dsModLevelDrop p,
+.dsMinLevelDrop p,
+.dsInsigLevelDrop p {
+  padding-left: 5px;
 }
 
 /* Filter Dropdown Container */
